server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pymongo==4.6.1
mongoengine==0.27.0
python-decouple==3.8
django-cors-headers==4.3.1
Pillow>=10.2.0
cloudinary==1.41.0

